+++ lib/react_email_rails/render_modes/persistent/server.rb
@@ -75,8 +75,6 @@ class ReactEmailRails::RenderModes::Persistent::Server
       }.tap do |body|
         body[:html] = response["html"] if response.key?("html")
         body[:text] = response["text"] if response.key?("text")
-        body[:warnings] = response["warnings"] if response.key?("warnings")
-        body[:document] = response["document"] if response.key?("document")
       end,
     ))
   rescue JSON::ParserError => e